+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Nessie Change Log
 
+## Next
+
+- Added support for custom seed and migration templates
+
 ## Version 2.0.3 - 2021-10-30
 
 - Deno v1.15.3